♻️ refactor: reorganize CI workflows and test directory structure
- Add ci-typescript.yml with categorized test execution:
  - PR: check + unit tests only (fast feedback)
  - Merge: check + unit + integration tests
  - Manual: all tests including optional e2e

- Add build.yml for tag-triggered multi-platform builds:
  - Linux, macOS, Windows binaries
  - Triggered on v* tags only
  - Artifacts stored for 90 days

- Reorganize tests directory:
  - tests/unit/: pure function tests (colors, commands, mod, auto-tag)
  - tests/integration/: module integration tests (cli, github-project, etc.)
  - tests/e2e/: placeholder for external API tests

- Update deno.json with new test tasks:
  - test:unit, test:integration, test:e2e

- Remove old ci.yml (replaced by ci-typescript.yml and build.yml)

Closes #24
