+++ .github/workflows/build.yml
@@ -1,0 +1,61 @@
+# Build Workflow
+# Triggered on tag push (v*) for multi-platform binary compilation
+#
+# This workflow compiles the CLI tool for multiple platforms:
+# - Linux (x86_64)
+# - macOS (x86_64)
+# - Windows (x86_64)
+#
+# Artifacts are uploaded for download but not released automatically.
+# Use a separate release workflow or manual release for GitHub Releases.
+
+name: Build
+
+on:
+  push:
+    tags:
+      - 'v*'
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+    name: Build (${{ matrix.os }})
+    runs-on: ${{ matrix.os }}
+
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        include:
+          - os: ubuntu-latest
+            target: x86_64-unknown-linux-gnu
+            output: hewg
+            artifact_name: hewg-linux-x86_64
+          - os: macos-latest
+            target: x86_64-apple-darwin
+            output: hewg
+            artifact_name: hewg-macos-x86_64
+          - os: windows-latest
+            target: x86_64-pc-windows-msvc
+            output: hewg.exe
+            artifact_name: hewg-windows-x86_64
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Deno
+        uses: denoland/setup-deno@v2
+        with:
+          deno-version: v2.x
+
+      - name: Compile
+        run: deno compile --allow-read --allow-env --allow-net --target ${{ matrix.target }} --output dist/${{ matrix.output }} src/main.ts
+
+      - name: Upload artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: ${{ matrix.artifact_name }}
+          path: dist/${{ matrix.output }}
+          retention-days: 90
