+++ .github/workflows/ci-typescript.yml
@@ -1,0 +1,190 @@
+# TypeScript/Deno CI Workflow
+# Generated by: hewg setup-actions --ci
+#
+# This workflow provides comprehensive CI for TypeScript/Deno projects:
+# - Code quality checks (formatter, linter, type checker)
+# - Categorized test execution (unit, integration, e2e)
+# - Coverage reporting via Codecov
+#
+# Test execution timing:
+# - PR: check + unit tests
+# - Merge: check + unit + integration tests
+# - Manual: all tests (e2e optional)
+#
+# Test directory structure:
+# tests/
+# ├── unit/           # Fast, isolated tests (every PR)
+# ├── integration/    # Module integration tests (after merge)
+# └── e2e/            # External API tests (manual only)
+
+name: CI
+
+on:
+  push:
+    branches: [main, develop]
+  pull_request:
+    branches: [main, develop]
+  workflow_dispatch:
+    inputs:
+      run_e2e:
+        description: 'Run E2E tests (includes external API calls)'
+        type: boolean
+        default: false
+
+permissions:
+  contents: read
+
+jobs:
+  check:
+    name: Check
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Deno
+        uses: denoland/setup-deno@v2
+        with:
+          deno-version: v2.x
+
+      - name: Verify formatting
+        run: deno fmt --check
+
+      - name: Run linter
+        run: deno lint
+
+      - name: Type check
+        run: deno check src/**/*.ts
+
+  test-unit:
+    name: Unit Tests
+    runs-on: ubuntu-latest
+    needs: check
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Deno
+        uses: denoland/setup-deno@v2
+        with:
+          deno-version: v2.x
+
+      - name: Run unit tests
+        run: deno test tests/unit/ --allow-read --allow-env --coverage=coverage/unit/
+
+      - name: Generate coverage report
+        run: deno coverage coverage/unit/ --lcov > coverage/unit.lcov
+
+      - name: Upload unit coverage artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: coverage-unit
+          path: coverage/unit.lcov
+          retention-days: 1
+
+  test-integration:
+    name: Integration Tests
+    runs-on: ubuntu-latest
+    needs: test-unit
+    # Run only on merge to main/develop or manual dispatch
+    if: github.event_name == 'push' || github.event_name == 'workflow_dispatch'
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Deno
+        uses: denoland/setup-deno@v2
+        with:
+          deno-version: v2.x
+
+      - name: Run integration tests
+        run: deno test tests/integration/ --allow-read --allow-env --allow-net --coverage=coverage/integration/
+
+      - name: Generate coverage report
+        run: deno coverage coverage/integration/ --lcov > coverage/integration.lcov
+
+      - name: Upload integration coverage artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: coverage-integration
+          path: coverage/integration.lcov
+          retention-days: 1
+
+  test-e2e:
+    name: E2E Tests
+    runs-on: ubuntu-latest
+    needs: check
+    # Run only on manual dispatch with explicit e2e flag
+    if: github.event_name == 'workflow_dispatch' && inputs.run_e2e == true
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Deno
+        uses: denoland/setup-deno@v2
+        with:
+          deno-version: v2.x
+
+      - name: Run E2E tests
+        run: deno test tests/e2e/ --allow-read --allow-env --allow-net
+
+  coverage:
+    name: Upload Coverage
+    runs-on: ubuntu-latest
+    needs: [test-unit, test-integration]
+    # Run only when integration tests run (ensures both coverages exist)
+    if: github.event_name == 'push' || github.event_name == 'workflow_dispatch'
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Download unit coverage
+        uses: actions/download-artifact@v4
+        with:
+          name: coverage-unit
+          path: coverage/
+
+      - name: Download integration coverage
+        uses: actions/download-artifact@v4
+        with:
+          name: coverage-integration
+          path: coverage/
+
+      - name: Merge coverage reports
+        run: |
+          cat coverage/unit.lcov coverage/integration.lcov > coverage/lcov.info
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v4
+        with:
+          files: coverage/lcov.info
+          fail_ci_if_error: false
+
+  # Coverage upload for PR (unit tests only)
+  coverage-pr:
+    name: Upload Coverage (PR)
+    runs-on: ubuntu-latest
+    needs: test-unit
+    # Run only on PR (when integration tests don't run)
+    if: github.event_name == 'pull_request'
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Download unit coverage
+        uses: actions/download-artifact@v4
+        with:
+          name: coverage-unit
+          path: coverage/
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v4
+        with:
+          files: coverage/unit.lcov
+          fail_ci_if_error: false
