--- .github/workflows/ci.yml
@@ -1,96 +1,0 @@
-name: CI
-
-on:
-  push:
-    branches: [main, develop]
-  pull_request:
-    branches: [main, develop]
-
-permissions:
-  contents: read
-
-jobs:
-  check:
-    name: Check
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Setup Deno
-        uses: denoland/setup-deno@v2
-        with:
-          deno-version: v2.x
-
-      - name: Verify formatting
-        run: deno fmt --check
-
-      - name: Run linter
-        run: deno lint
-
-      - name: Type check
-        run: deno check src/**/*.ts tests/**/*.ts
-
-  test:
-    name: Test
-    runs-on: ubuntu-latest
-    needs: check
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Setup Deno
-        uses: denoland/setup-deno@v2
-        with:
-          deno-version: v2.x
-
-      - name: Run tests
-        run: deno test --allow-read --allow-env --coverage=coverage/
-
-      - name: Generate coverage report
-        run: deno coverage coverage/ --lcov > coverage/lcov.info
-
-      - name: Upload coverage
-        uses: codecov/codecov-action@v4
-        with:
-          files: coverage/lcov.info
-          fail_ci_if_error: false
-
-  build:
-    name: Build
-    runs-on: ${{ matrix.os }}
-    needs: test
-
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        include:
-          - os: ubuntu-latest
-            target: x86_64-unknown-linux-gnu
-            output: hewg
-          - os: macos-latest
-            target: x86_64-apple-darwin
-            output: hewg
-          - os: windows-latest
-            target: x86_64-pc-windows-msvc
-            output: hewg.exe
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Setup Deno
-        uses: denoland/setup-deno@v2
-        with:
-          deno-version: v2.x
-
-      - name: Compile
-        run: deno compile --allow-read --allow-env --allow-net --target ${{ matrix.target }} --output dist/${{ matrix.output }} src/main.ts
-
-      - name: Upload artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: hewg-${{ matrix.target }}
-          path: dist/${{ matrix.output }}
