+++ deno.json
@@ -10,6 +10,9 @@
     "dev": "deno run --watch src/main.ts",
     "start": "deno run src/main.ts",
     "test": "deno test --allow-read --allow-env",
+    "test:unit": "deno test tests/unit/ --allow-read --allow-env",
+    "test:integration": "deno test tests/integration/ --allow-read --allow-env --allow-net",
+    "test:e2e": "deno test tests/e2e/ --allow-read --allow-env --allow-net",
     "test:watch": "deno test --allow-read --allow-env --watch",
     "test:coverage": "deno test --allow-read --allow-env --coverage=coverage/",
     "coverage": "deno coverage coverage/ --lcov > coverage/lcov.info",
