+++ tests/e2e/.gitkeep
@@ -1,0 +1,2 @@
+# Placeholder for E2E tests directory
+# E2E tests requiring external API calls should be placed here
